+++ src/test/java/com/zepben/ewb/datastores/energy/SqliteByDateBlobStoreProviderTest.java
@@ -11,6 +11,7 @@ package com.zepben.ewb.datastores.energy;
 import com.zepben.blobstore.BlobStoreException;
 import com.zepben.blobstore.sqlite.SqliteBlobStore;
 import com.zepben.evolve.database.paths.EwbDataFilePaths;
+import com.zepben.evolve.database.paths.LocalEwbDataFilePaths;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -41,7 +42,7 @@ public class SqliteByDateBlobStoreProviderTest {
     @BeforeEach
     public void before(@TempDir Path tempDir) {
         folder = tempDir.toFile();
-        EwbDataFilePaths ewbFilePaths = new EwbDataFilePaths(tempDir.toString());
+        EwbDataFilePaths ewbFilePaths = new LocalEwbDataFilePaths(tempDir.toString());
         provider = new SqliteByDateBlobStoreProvider(ewbFilePaths);
     }